Updated StarCore and Java

+++ src/main/java/com/thenexusreborn/survivalgames/mutations/Mutation.java
@@ -75,7 +75,7 @@ public abstract class Mutation {
         this.countdownTimer.addRepeatingCallback(new MutationCountdownCallback(game, this), TimeUnit.SECONDS, 1);
         this.countdownTimer.setEndCondition(new MutationEndCondition(this, game));
         this.countdownTimer.setLength(game.getSettings().getMutationSpawnDelay() * 1000L + 50);
-        long seconds = TimeUnit.MILLISECONDS.toSeconds(this.countdownTimer.getTime());
+        long seconds = (long) TimeUnit.MILLISECONDS.toSeconds(this.countdownTimer.getTime());
         p.sendMessage(ColorHandler.getInstance().color("&6&l>> &eYou will mutate as a(n) " + getType().getDisplayName() + "!"));
         p.sendMessage(ColorHandler.getInstance().color("&6&l>> &eYou will be mutated in &l" + seconds + " Seconds&e."));
         Player t = Bukkit.getPlayer(this.target);
